+++ plugin/item/stat40/stat40_repair.c
@@ -85,7 +85,8 @@ errno_t stat40_check_struct(reiser4_place_t *place, repair_hint_t *hint) {
 	if (res) {
 		fsck_mess("Node (%llu), item (%u), [%s]: does "
 			  "not look like a valid stat data.", 
-			  place_blknr(place), place->pos.item,
+			  (unsigned long long)place_blknr(place),
+			  place->pos.item,
 			  print_key(stat40_core, &place->key));
 		
 		return RE_FATAL;
@@ -94,9 +95,11 @@ errno_t stat40_check_struct(reiser4_place_t *place, repair_hint_t *hint) {
 	if (stat.len + hint->len < place->len) {
 		fsck_mess("Node (%llu), item (%u), [%s]: item has the "
 			  "wrong length (%u). Should be (%llu).%s",
-			  place_blknr(place), place->pos.item, 
+			  (unsigned long long)place_blknr(place),
+			  place->pos.item,
 			  print_key(stat40_core, &place->key),
-			  place->len, stat.len + hint->len, 
+			  place->len,
+			  (unsigned long long)(stat.len + hint->len),
 			  hint->mode == RM_BUILD ? " Fixed." : "");
 		
 		if (hint->mode != RM_BUILD)
@@ -110,11 +113,14 @@ errno_t stat40_check_struct(reiser4_place_t *place, repair_hint_t *hint) {
 	if (stat.extmask != stat.goodmask) {
 		fsck_mess("Node (%llu), item (%u), [%s]: item has the "
 			  "wrong extention mask (%llu). Should be (%llu)."
-			  "%s", place_blknr(place), place->pos.item,
-			  print_key(stat40_core, &place->key), stat.extmask,
-			  stat.goodmask, hint->mode == RM_CHECK ? "" : 
-			  " Fixed.");
-		
+			  "%s",
+			  (unsigned long long)place_blknr(place),
+			  place->pos.item,
+			  print_key(stat40_core, &place->key),
+			  (unsigned long long)stat.extmask,
+			  (unsigned long long)stat.goodmask,
+			  hint->mode == RM_CHECK ? "" : " Fixed.");
+
 		if (hint->mode == RM_CHECK)
 			return RE_FIXABLE;
